fix float precisio and loss backward

+++ utils/glove_loader.py
@@ -9,7 +9,7 @@ def load_glove_model(File, vocab):
             split_line = line.split()
             word = split_line[0]
             if word in vocab:
-              embedding = np.array(split_line[1:], dtype=np.float64)
+              embedding = np.array(split_line[1:], dtype=np.float32)
               glove_model[word] = embedding
           except ValueError :
             continue
